Reomved the about page file. created separate components for the 2 events. balanced button and title spacing.
